N/A: Moved DBContext and DBContextFactory to resources/base.py
from mfit/app.py
